feat: main service rename packages
