+++ ewm-service/src/main/java/ru/practicum/dto/event/EventShortDto.java
@@ -1,8 +1,8 @@
-package ru.practicum.dto.events;
+package ru.practicum.dto.event;
 
 import lombok.Data;
-import ru.practicum.dto.categories.CategoryDto;
-import ru.practicum.dto.users.UserShortDto;
+import ru.practicum.dto.category.CategoryDto;
+import ru.practicum.dto.user.UserShortDto;
 
 @Data
 public class EventShortDto {
