+++ ewm-service/src/main/java/ru/practicum/dto/event/NewEventDto.java
@@ -1,14 +1,12 @@
-package ru.practicum.dto.events;
+package ru.practicum.dto.event;
 
 import jakarta.validation.constraints.Positive;
 import lombok.Data;
 import org.hibernate.validator.constraints.Length;
-import ru.practicum.dto.events.enums.StateActionUser;
-import ru.practicum.dto.locations.LocationDto;
-import ru.practicum.dto.validation.enums.EnumValidator;
+import ru.practicum.dto.location.LocationDto;
 
 @Data
-public class UpdateEventUserRequest {
+public class NewEventDto {
     @Length(min = 20, max = 2000)
     private String annotation;
     @Positive
@@ -17,11 +15,9 @@ public class UpdateEventUserRequest {
     private String description;
     private String eventDate;
     private LocationDto location;
-    private Boolean paid;
-    private Integer participantLimit;
-    private Boolean requestModeration;
-    @EnumValidator(enumClazz = StateActionUser.class)
-    private String stateAction;
+    private boolean paid = false;
+    private int participantLimit = 0;
+    private boolean requestModeration = true;
     @Length(min = 3, max = 120)
     private String title;
 }
