+++ ewm-service/src/main/java/ru/practicum/dto/event/UpdateEventUserRequest.java
@@ -1,14 +1,14 @@
-package ru.practicum.dto.events;
+package ru.practicum.dto.event;
 
 import jakarta.validation.constraints.Positive;
 import lombok.Data;
 import org.hibernate.validator.constraints.Length;
-import ru.practicum.dto.events.enums.StateActionAdmin;
-import ru.practicum.dto.locations.LocationDto;
+import ru.practicum.dto.event.enums.StateActionUser;
+import ru.practicum.dto.location.LocationDto;
 import ru.practicum.dto.validation.enums.EnumValidator;
 
 @Data
-public class UpdateEventAdminRequest {
+public class UpdateEventUserRequest {
     @Length(min = 20, max = 2000)
     private String annotation;
     @Positive
@@ -20,7 +20,7 @@ public class UpdateEventAdminRequest {
     private Boolean paid;
     private Integer participantLimit;
     private Boolean requestModeration;
-    @EnumValidator(enumClazz = StateActionAdmin.class)
+    @EnumValidator(enumClazz = StateActionUser.class)
     private String stateAction;
     @Length(min = 3, max = 120)
     private String title;
